Services/audit_event/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

msal
azure-mgmt-web
azure-mgmt-authorization
azure-functions
azure-functions-durable
azure-servicebus
azure-mgmt-resource
azure-mgmt-storage
azure-mgmt-containerservice
azure-mgmt-containerinstance
azure-mgmt-subscription
azure-mgmt-compute
azure-mgmt-network
azure-keyvault-secrets
azure-identity
azure-mgmt-sql
azure-storage-blob